April 24 Progress Report
Hello Profressor, for this progress report I aimed to write code for Javascript and JSON. I learned how to program a bit better in JavaScript by following this quick online course. It provided me with the essentials and gave me a good recap on the language. I've learned how to print out a variable within a string using $(). I also got a better understanding of the various types included in this language including objects, booleans, nulls, symbols

This Javascript program that I created is really simple. Basically, it's given some constant string values inputted as numbers with included symbols like (), -, specific formats. Then, in the console, it will perform multiple tests to find the numbers that have parenthesis, dashes, parenthesis and/or dashes, and find a specific number formate like (###) ###-#### The numbers that follow the correct statement get outputted as true. Those that are not, will output as false.

As for the JSON, I got a bit confused on this matter, but after doing a bit of research I figured out it was a data representation format. Sort of like XML. In the program that I created some data which we're numbers. And then I outputted the numbers to the console using JSON.parse.